Simplify basename implementation
