+++ packages/php-wasm/util/src/lib/index.ts
@@ -1,8 +1,7 @@
 import Semaphore from './semaphore';
 export { Semaphore };
 export type { SemaphoreOptions } from './semaphore';
-export { basename } from './basename';
-export { joinPaths } from './join-paths';
+export { joinPaths, basename } from './paths';
 export { createSpawnHandler } from './create-spawn-handler';
 
 export * from './php-vars';
